add crude is file in use check in deleteone

+++ src/DatabaseConnection.ts
@@ -3622,6 +3622,18 @@ export class DatabaseConnection extends events.EventEmitter {
                             throw new Error(`[${user.name}] Access denied, missing delete permission`);
                         }
                     }
+                    let cursor = this.db.collection("agents").find({ "fileid": id });
+                    let locked = await cursor.hasNext();
+                    if(locked) {
+                        let item = await cursor.next();
+                        throw new Error("Cannot delete file, it is in use by an agent " + item._id + " " + item.name);
+                    }
+                    cursor = this.db.collection("workitems").find({ "files._id": id });
+                    locked = await cursor.hasNext();
+                    if(locked) {
+                        let item = await cursor.next();
+                        throw new Error("Cannot delete file, it is in use by a workitem " + item._id + " " + item.name);
+                    }
 
                     const ot_end = Logger.otel.startTimer();
                     await this._DeleteFile(id, collectionname);
